Move discover row section to separate file
